+++ src/track.c
@@ -145,10 +145,16 @@ Track_starred(Track * self, PyObject *args, PyObject *kwds)
 }
 
 static PyObject *
-Track_availability(Track* self)
+Track_availability(Track *self)
 {
     return Py_BuildValue("i",
                          sp_track_get_availability(g_session, self->_track));
+}
+
+static PyObject *
+Track_is_local(Track *self)
+{
+    return PyBool_FromLong(sp_track_is_local(g_session, self->_track));
 }
 
 static PyMethodDef Track_methods[] = {
@@ -160,6 +166,10 @@ static PyMethodDef Track_methods[] = {
      (PyCFunction)Track_is_loaded,
      METH_NOARGS,
      "Get load status for this track. If the track is not loaded yet, all other functions operating on the track return default values."},
+    {"is_local",
+     (PyCFunction)Track_is_local,
+     METH_NOARGS,
+     "Returns wether this track is local, ie. created with the API."},
     {"artists",
      (PyCFunction)Track_artists,
      METH_VARARGS,